Use a session scoped fixture
